complete function(required changes)
